+++ src/com/intellij/idea/plugin/hybris/common/utils/HybrisIcons.kt
@@ -174,6 +174,8 @@ object HybrisIcons {
         val Y_COLUMN_ALL = getIcon("/icons/flexibleSearch/star.svg")
         val TABLE_SUFFIX = AllIcons.General.Filter
         val TABLE_ALIAS_SEPARATOR = getIcon("/icons/flexibleSearch/separator.svg")
+        val HIDE_PARAMETERS_PANEL = getIcon("/icons/flexibleSearch/actions/fxsEditorOnly.svg")
+        val SHOW_PARAMETERS_PANEL = getIcon("/icons/flexibleSearch/actions/fxsEditorWithParametersPanel.svg")
 
         object Actions {
             val GROUP = getIcon("/icons/flexibleSearch/actions/group.svg")