+++ src/com/intellij/idea/plugin/hybris/tools/logging/actions/LoggerActions.kt
@@ -45,7 +45,7 @@ abstract class AbstractLoggerAction(private val logLevel: String, val icon: Icon
         val project = e.project ?: return
 
         val dataContext = e.dataContext
-        val logIdentifier = dataContext.getData(HybrisConstants.LOGGER_IDENTIFIER_DATA_CONTEXT_KEY)
+        val logIdentifier = dataContext.getData(HybrisConstants.KEY_LOGGER_IDENTIFIER_DATA_CONTEXT)
 
         if (logIdentifier == null) {
             notify(